adds graphql-server crate
