+++ server-rust/Cargo.toml
@@ -3,4 +3,5 @@
 members = [
     "data-access",
     "utilities",
+    "graphql-server",
 ]
