--- server-rust/data-access/src/main.rs
@@ -1,25 +1,0 @@
-use std::error::Error;
-
-use mongodb::{
-    options::{ClientOptions, ResolverConfig},
-    Client,
-};
-
-pub mod api;
-pub mod error;
-pub mod model;
-
-#[tokio::main]
-async fn main() -> Result<(), Box<dyn Error>> {
-    // Load the MongoDB connection string from an environment variable:
-    let client_uri =
-        std::env::var("MONGODB_URI").expect("You must set the MONGODB_URI environment var!");
-
-    // A Client is needed to connect to MongoDB:
-    // An extra line of code to work around a DNS issue on Windows:
-    let options =
-        ClientOptions::parse_with_resolver_config(&client_uri, ResolverConfig::cloudflare())
-            .await?;
-    let client = Client::with_options(options)?;
-    Ok(())
-}
