+++ server-rust/graphql-server/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "graphql-server"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+axum = "0.6.13"
+juniper = "0.15.11"
+juniper_warp = "0.7.0"
+serde = "1.0.160"
+tokio = "1.27.0"
