+++ server-rust/graphql-server/src/main.rs
@@ -1,0 +1,6 @@
+use std::error::Error;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn Error>> {
+    Ok(())
+}
